Adding GraphicsEx example
Adding a GraphicsEx example, but I am using FlashDevelop, so the project
structure is a bit different. Naturally, the examples should all have
the same structure, but I can't achieve that.

Sample includes a version of the Ripple UV Vertex shader together with a
TwoTextureVertexColorFragmentShader, creating an interesting effect by
placing the same glow texture ontop of eachother, animating it at
different speeds. Outside of the scope for the GraphicsEx demo, perhaps.