Adding the FastStroke class and Benchmark
The FastStroke is a class that is optimized for speed. The method
setCapacity has to be called with the amount of vertices that will be
added. The class generates a 4 vertex rectangle for each line segments,
meaning that thicker lines will look bad in the joints. Use for thin
lines where a large number of lines is needed