Fixing Lost Context bug in FastStroke
FastStroke needed to know if we lost context. I added
onGraphicLostContext to the base class Graphic to give the FastStroke
awareness of this happening without having to add a second event
listener for the same event. I did not want to break the "private"-ness
of onContextCreated in the base class, I think this is a decent middle
ground